pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tasklist"
version = "0.1.0"
description = "Django Task List Application"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "djangorestframework-simplejwt>=5.3.0",
]

[tool.setuptools.packages.find]
exclude = ["static*", "tests*"]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "playwright>=1.40.0",
    "pytest-xvfb>=3.0.0",
    "locust>=2.17.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tasklist_project.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing --strict-markers"

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
    "*/test_*.py",
]
